Move AudioBufferSourceNode tests to WPT
These tests basically pass on Firefox.  If they don't, it's because
Firefox is incorrect and issues have been filed for these.

During the testing, two mistakes in the tests were found.

One is that we incorrectly handled some promises and the task wasn't
waiting for everything to finish.

The second was an incorrect assumption that calling start() would
work.  Because of test failures in Firefox, this assumption is
incorrect.

Updated audit.js to latest Chrome version (to get loadFileFromURL) and
add wpt lint exception for a console message in audit.js.

Bug: 745778
Change-Id: I5f9ee01498678db0feb57051f682772ca7bd9fbb